add restrictions

+++ base/views.py
@@ -1,7 +1,8 @@
-from email import message
 from django.shortcuts import render, redirect
 from django.contrib import messages
+from django.http import HttpResponse
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.db.models import Q
 from .models import Room, Topic
@@ -10,6 +11,10 @@ from .forms import RoomForm
 
 def login_page(request):
     context = {}
+
+    if request.user.is_authenticated:
+        return redirect('home')
+    
     if request.method == "POST":
         username = request.POST.get("username")
         password = request.POST.get("password")
@@ -27,9 +32,11 @@ def login_page(request):
 
     return render(request, 'base/login_register.html', context)
 
+
 def logout_user(request):
     logout(request)
     return redirect('home')
+
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') else ''
@@ -41,11 +48,13 @@ def home(request):
     context = {"rooms": rooms, "topics": topics, "room_count": rooms.count()}
     return render(request, 'base/home.html', context)
 
+
 def room(request, pk):
     context = {"room": None}
     context["room"] = Room.objects.get(id=pk)
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def create_room(request):
     if request.method == "POST":
         form = RoomForm(request.POST)
@@ -56,9 +65,14 @@ def create_room(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not authorized to perform the action!")
+
     context = {'form': form}
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -66,9 +80,12 @@ def update_room(request, pk):
         return redirect('home')
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
     context = {'obj': room}
+    if request.user != room.host:
+        return HttpResponse("You are not authorized to perform the action!")
     if request.method == "POST":
         room.delete()
         return redirect('home')
